Updated frame reading and queue termination
Changed the frame reading by doing a grab of all frames, but only retrieve the ones we want. This makes the frame extraction much faster
Once there are no more frames, an "ending frame" is placed, that has a frame number = -1. The other threads pick it up and know that the stream is over